refactoring for multi-dumper support

+++ packages/oops.staticdump/branches/book-dumpers-removal/oops/staticdump/controlpanel.py
@@ -1,33 +1,56 @@
 import os
 import shutil
-import transaction
 
 from collective.transmogrifier.interfaces import ITransmogrifier
+
 from plone.app.controlpanel.form import ControlPanelForm
 from plone.fieldsets.form import FieldsetsEditForm
 from zope.app.component.hooks import getSite
 from zope.formlib import form
+
+from zope.interface import directlyProvides
 from zope.interface import Interface
 from zope.interface import implements
-from zope.schema import TextLine
+
+from zope.schema import Choice #TextLine
+from zope.schema.interfaces import IVocabularyFactory
+from zope.schema.vocabulary import SimpleVocabulary, SimpleTerm
 
 from Products.CMFCore.utils import getToolByName
 from Products.CMFDefault.formlib.schema import ProxyFieldProperty
 from Products.CMFDefault.formlib.schema import SchemaAdapterBase
 from Products.Five.browser.pagetemplatefile import ZopeTwoPageTemplateFile
 
+# fixme: remove old fields if no more needed
+
+def DumpersVocabularyFactory(context):
+    """Vocabulary factory for dumpers ids
+    """
+    portal = getSite()
+    tool = getToolByName(portal, 'portal_dumper')
+    terms = [SimpleTerm(o.getId(), title=o.title) for o in tool.objectValues()]
+    return SimpleVocabulary(terms)
+
+directlyProvides(DumpersVocabularyFactory, IVocabularyFactory)
+
 
 class IDumpPanelSchema(Interface):
 
-    filesystem_target = TextLine(
-        title=u'Filesystem target',
-        description=u"Path on server filesystem",
+    dumper = Choice(
+        title=u'Dumpers',
+        description=u'Available dumpers',
+        vocabulary=u'oops.staticdump.dumpers'
     )
 
-    html_base = TextLine(
-        title=u'HTML base',
-        description=u"Base attribute value used in html files",
-    )
+    #filesystem_target = TextLine(
+        #title=u'Filesystem target',
+        #description=u"Path on server filesystem",
+    #)
+
+    #html_base = TextLine(
+        #title=u'HTML base',
+        #description=u"Base attribute value used in html files",
+    #)
 
 
 class DumpPanelAdapter(SchemaAdapterBase):
@@ -37,11 +60,13 @@ class DumpPanelAdapter(SchemaAdapterBase):
     def __init__(self, context):
         super(DumpPanelAdapter, self).__init__(context)
         self.portal = getSite()
-        pprop = getToolByName(self.portal, 'portal_properties')
-        self.context = pprop.dumper_properties
+        dumper_tool = getToolByName(self.portal, 'portal_dumper')
+        self.context = dumper_tool
 
-    filesystem_target = ProxyFieldProperty(IDumpPanelSchema['filesystem_target'])
-    html_base = ProxyFieldProperty(IDumpPanelSchema['html_base'])
+    dumper = ProxyFieldProperty(IDumpPanelSchema['dumper'])
+
+    #filesystem_target = ProxyFieldProperty(IDumpPanelSchema['filesystem_target'])
+    #html_base = ProxyFieldProperty(IDumpPanelSchema['html_base'])
 
 
 class DumpPanel(ControlPanelForm):
@@ -54,11 +79,11 @@ class DumpPanel(ControlPanelForm):
     description = u"Update static version of this site"
     form_name = u'Dump'
 
-    def setUpWidgets(self, ignore_request=False):
-        # this is the only way to call the super methods
-        FieldsetsEditForm.setUpWidgets(self, ignore_request)
-        self.widgets['filesystem_target'].displayWidth = 50
-        self.widgets['html_base'].displayWidth = 50
+    #def setUpWidgets(self, ignore_request=False):
+        ## this is the only way to call the super methods
+        #FieldsetsEditForm.setUpWidgets(self, ignore_request)
+        ##self.widgets['filesystem_target'].displayWidth = 50
+        ##self.widgets['html_base'].displayWidth = 50
 
 
     def __call__(self):
@@ -73,24 +98,26 @@ class DumpPanel(ControlPanelForm):
 
         return self.render()
 
-    def _applyChanges(self, data):
-        # save the base attribute always with an ending '/'
-        html_base = data.get('html_base', '')
-        if not html_base.endswith('/'):
-            data['html_base'] = html_base + '/'
+    #def _applyChanges(self, data):
+        ## save the base attribute always with an ending '/'
+        #html_base = data.get('html_base', '')
+        #if not html_base.endswith('/'):
+            #data['html_base'] = html_base + '/'
 
-        form.applyChanges(self.context, self.form_fields, data, self.adapters)
+        #form.applyChanges(self.context, self.form_fields, data, self.adapters)
 
     @form.action(u'Save', name=u'save')
     def handle_save(self, action, data):
-        self._applyChanges(data)
+        #self._applyChanges(data)
+        tool = getToolByName(self.context, 'portal_dumper')
+        tool.setDumper(data['dumper'])
         self.status = u'Save completed.'
 
     @form.action(u'Save and dump now', name=u'dump')
     def handle_dump(self, action, data):
-        self._applyChanges(data)
-        props = getToolByName(self.context,
-                             'portal_properties').dumper_properties
+        #self._applyChanges(data)
+        tool = getToolByName(self.context, 'portal_dumper')
+        tool.setDumper(data['dumper'])
 
         if dump_is() == running:
             self.status = 'Dump running.'
@@ -98,17 +125,20 @@ class DumpPanel(ControlPanelForm):
             dump_is(running)
 
             # do dump
-            destination = data.get('filesystem_target')
+            #destination = data.get('filesystem_target')
+            destination = tool.getDumperProperty('filesystem_target')
             tmp = destination + '_tmp'
 
             transmogrifier_overrides = {
                 'destination': tmp,
-                'static_base': data.get('html_base')
+                #'static_base': data.get('html_base')
+                'static_base':tool.getDumperProperty('html_base')
             }
 
             transmogrifier = ITransmogrifier(self.context)
-            configuration_name = props.getProperty('dump_configuration_name',
-                                                   'dump_sample')
+            configuration_name = tool.getDumperProperty(
+                                    'dump_configuration_name',
+                                    'dump_sample')
             try:
                 transmogrifier(configuration_name,
                                transmogrifier = transmogrifier_overrides)
@@ -122,21 +152,17 @@ class DumpPanel(ControlPanelForm):
 
 
 def backup_and_switch(destination, tmp):
-    old = destination + '_old'
-    if os.path.exists(old):
-        shutil.rmtree(old)
-    shutil.move(destination, old)
-    shutil.move(tmp, destination)
+    if os.path.exists(destination):
+        old = destination + '_old'
+        if os.path.exists(old):
+            shutil.rmtree(old)
+        shutil.move(destination, old)
+        shutil.move(tmp, destination)
 
 
 running = True
 not_running = False
 def dump_is(status=None):
     portal = getSite()
-    dp = getToolByName(portal, 'portal_properties').dumper_properties
-    if status is None:
-        return  dp.getProperty('running')
-    else:
-        dp._updateProperty('running', status)
-        transaction.commit()
-        return status
+    tool = getToolByName(portal, 'portal_dumper')
+    return tool.status(status)
